@@ -9,35 +9,35 @@
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e",
-    "postinstall": "ngcc"
+    "e2e": "ng e2e"
   },
   "private": true,
   "browserslist": [],
   "dependencies": {
-    "@angular/animations": "9.1.13",
-    "@angular/cdk": "9.2.4",
-    "@angular/cli": "9.1.15",
-    "@angular/common": "9.1.13",
-    "@angular/compiler": "9.1.13",
-    "@angular/core": "9.1.13",
-    "@angular/forms": "9.1.13",
-    "@angular/localize": "9.1.13",
-    "@angular/material": "9.2.4",
-    "@angular/platform-browser": "9.1.13",
-    "@angular/platform-browser-dynamic": "9.1.13",
-    "@angular/platform-server": "9.1.13",
-    "@angular/router": "9.1.13",
-    "@auth0/auth0-angular": "1.10.1",
-    "@ng-bootstrap/ng-bootstrap": "6.1.0",
+    "@angular/animations": "17.3.8",
+    "@angular/cdk": "17.3.8",
+    "@angular/cli": "17.3.7",
+    "@angular/common": "17.3.8",
+    "@angular/compiler": "17.3.8",
+    "@angular/core": "17.3.8",
+    "@angular/forms": "17.3.8",
+    "@angular/localize": "17.3.8",
+    "@angular/material": "^17.3.8",
+    "@angular/platform-browser": "17.3.8",
+    "@angular/platform-browser-dynamic": "17.3.8",
+    "@angular/platform-server": "17.3.8",
+    "@angular/router": "17.3.8",
+    "@auth0/auth0-angular": "2.2.3",
+    "@ng-bootstrap/ng-bootstrap": "^16.0.0",
+    "@popperjs/core": "^2.11.8",
     "@types/mathjax": "0.0.35",
     "@types/mathjs": "0.0.38",
+    "bootstrap": "^5.3.2",
     "compodoc": "0.0.41",
     "core-js": "3.2.1",
     "css-loader": "3.2.0",
     "d3": "5.11.0",
     "file-loader": "4.2.0",
-    "hammerjs": "2.0.8",
     "jasmine": "3.5.0",
     "jasminewd2": "2.2.0",
     "jquery": "3.4.1",
@@ -45,8 +45,8 @@
     "mathjax": "2.7.1",
     "mathjs": "6.1.0",
     "ng-animate": "0.3.4",
-    "ngx-bootstrap": "6.1.0",
-    "ngx-logger": "4.0.3",
+    "ngx-bootstrap": "^12.0.0",
+    "ngx-logger": "^5.0.12",
     "npm": "6.11.2",
     "popper.js": "1.15.0",
     "rxjs": "6.6.7",
@@ -55,32 +55,32 @@
     "url-loader": "2.1.0",
     "webdriver-manager": "12.1.7",
     "webpack": "4.39.2",
-    "zone.js": "0.10.2"
+    "zone.js": "0.14.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.15",
-    "@angular/compiler-cli": "9.1.13",
-    "@angular/language-service": "9.1.13",
+    "@angular-devkit/build-angular": "^17.3.7",
+    "@angular/compiler-cli": "17.3.8",
+    "@angular/language-service": "17.3.8",
     "@types/d3": "^5.7.2",
     "@types/d3-hierarchy": "^1.1.6",
     "@types/jasmine": "3.4.0",
     "@types/jasminewd2": "^2.0.6",
     "@types/node": "^14.17.0",
-    "codelyzer": "^5.1.2",
+    "codelyzer": "6.0.2",
     "dockerlint": "^0.3.9",
     "jasmine-core": "^3.5.0",
     "jasmine-reporters": "^2.3.0",
     "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^4.2.0",
+    "karma": "^6.4.3",
     "karma-chrome-launcher": "^3.1.0",
     "karma-cli": "~2.0.0",
     "karma-jasmine": "^2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
     "log4js": "^5.1.0",
-    "protractor": "6.0.0",
+    "protractor": "7.0.0",
     "protractor-jasmine2-html-reporter": "0.0.7",
     "ts-node": "^8.3.0",
-    "tslint": "^5.19.0",
-    "typescript": "3.8.3"
+    "tslint": "^6.1.3",
+    "typescript": "5.4.5"
   }
 }
