@@ -2,7 +2,7 @@ import {Component, DoCheck, OnDestroy, OnInit, ViewChild} from '@angular/core';
 import {Subscription} from 'rxjs';
 import {ISUFactors} from '../../shared/model/ISUFactors';
 import {StudyService} from '../../shared/services/study.service';
-import {FormBuilder, FormGroup} from '@angular/forms';
+import {UntypedFormBuilder, UntypedFormGroup} from '@angular/forms';
 import {isNullOrUndefined} from 'util';
 import {ModalDismissReasons, NgbModal} from '@ng-bootstrap/ng-bootstrap';
 import {NavigationService} from '../../shared/services/navigation.service';
@@ -16,7 +16,7 @@ import {NGXLogger} from 'ngx-logger';
 export class HypothesisTheta0Component implements OnInit, OnDestroy {
 
   private _isuFactors: ISUFactors;
-  private _theta0Form: FormGroup;
+  private _theta0Form: UntypedFormGroup;
   private _isuFactorsSubscription: Subscription;
   private _visible: boolean;
   private _showHelpTextSubscription: Subscription;
@@ -25,7 +25,7 @@ export class HypothesisTheta0Component implements OnInit, OnDestroy {
   private helpTextModalReference: any;
   private _afterInit: boolean;
 
-  constructor(private _fb: FormBuilder,
+  constructor(private _fb: UntypedFormBuilder,
               private study_service: StudyService,
               private navigation_service: NavigationService,
               private modalService: NgbModal,
@@ -123,19 +123,19 @@ export class HypothesisTheta0Component implements OnInit, OnDestroy {
     this._isuFactorsSubscription = value;
   }
 
-  get theta0Form(): FormGroup {
+  get theta0Form(): UntypedFormGroup {
     return this._theta0Form;
   }
 
-  set theta0Form(value: FormGroup) {
+  set theta0Form(value: UntypedFormGroup) {
     this._theta0Form = value;
   }
 
-  get fb(): FormBuilder {
+  get fb(): UntypedFormBuilder {
     return this._fb;
   }
 
-  set fb(value: FormBuilder) {
+  set fb(value: UntypedFormBuilder) {
     this._fb = value;
   }
 
