@@ -7,7 +7,7 @@ import {PartialMatrix} from '../../shared/model/PartialMatrix';
 import {isNullOrUndefined} from 'util';
 import {Router} from '@angular/router';
 import {NGXLogger} from 'ngx-logger';
-import {FormBuilder, FormGroup} from '@angular/forms';
+import {UntypedFormBuilder, UntypedFormGroup} from '@angular/forms';
 import {NavigationService} from '../../shared/services/navigation.service';
 import {ModalDismissReasons, NgbModal} from '@ng-bootstrap/ng-bootstrap';
 import {ContrastMatrixService} from '../custom-contrast-matrix/contrast-matrix.service';
@@ -31,7 +31,7 @@ export class HypothesisWithinComponent implements OnInit, OnDestroy {
   private _isuFactors: ISUFactors;
   private _formErrors = constants.HYPOTHESIS_WITHIN_FORM_ERRORS;
   private _validationMessages = constants.HYPOTHESIS_WITHIN_VALIDATION_MESSAGES;
-  private _noColsForm: FormGroup;
+  private _noColsForm: UntypedFormGroup;
   private _maxCols: number;
   private _numCustomCols: number;
   private _contrast_matrix: PartialMatrix;
@@ -60,7 +60,7 @@ export class HypothesisWithinComponent implements OnInit, OnDestroy {
   constructor(private study_service: StudyService,
               private navigation_service: NavigationService,
               private contrast_matrix_service: ContrastMatrixService,
-              private fb: FormBuilder,
+              private fb: UntypedFormBuilder,
               private router: Router,
               private modalService: NgbModal,
               private log: NGXLogger) {
@@ -417,7 +417,7 @@ export class HypothesisWithinComponent implements OnInit, OnDestroy {
     return this._formErrors;
   }
 
-  get noRowsForm(): FormGroup {
+  get noRowsForm(): UntypedFormGroup {
     return this._noColsForm;
   }
 
