@@ -1,10 +1,10 @@
 import {Injectable} from '@angular/core';
-import {ActivatedRouteSnapshot, CanDeactivate, RouterStateSnapshot} from '@angular/router';
+import { ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import {BetweenIsuPredictorsComponent} from '../../study-form/between-isu-predictors/between-isu-predictors.component';
 import {Observable} from 'rxjs';
 
 @Injectable()
-export class CanDeactivatePredictorsGuardService implements CanDeactivate<BetweenIsuPredictorsComponent> {
+export class CanDeactivatePredictorsGuardService  {
   constructor() {}
 
   canDeactivate(component: BetweenIsuPredictorsComponent,
