@@ -1,12 +1,12 @@
 import { Injectable } from '@angular/core';
-import {ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot} from '@angular/router';
+import { ActivatedRouteSnapshot, Router, RouterStateSnapshot } from '@angular/router';
 import {StudyService} from './study.service';
 import {ISUFactors} from '../model/ISUFactors';
 import {Subscription} from 'rxjs';
 import {NGXLogger} from 'ngx-logger';
 
 @Injectable()
-export class MarginalMeansGuard implements CanActivate {
+export class MarginalMeansGuard  {
   private isuFactors: ISUFactors;
   private isuFactorsSubscription: Subscription;
 
