@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import {ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot} from '@angular/router';
+import { ActivatedRouteSnapshot, Router, RouterStateSnapshot } from '@angular/router';
 import {StudyService} from './study.service';
 import {Subscription} from 'rxjs';
 import {NGXLogger} from 'ngx-logger';
@@ -9,7 +9,7 @@ import {constants} from '../model/constants';
  * SolveForPower guard allows access to a route if and only if the currently loaded StudyDesign is solving for power.
  */
 @Injectable()
-export class SolveForPowerGuard implements CanActivate {
+export class SolveForPowerGuard  {
   private solveForPower: boolean;
   private solveForPowerSubscription: Subscription;
 
