@@ -9,8 +9,8 @@ export const environment = {
   name: 'dev',
   // serverLoggingUrl: 'http://localhost:5000/api/clientsidelog',
   // calculateUrl: 'http://localhost:5000/api/calculate',
-  serverLoggingUrl: 'https://glimmpse.samplesizeshop.org:5000/api/clientsidelog',
-  calculateUrl: 'https://glimmpse.samplesizeshop.org:5000/api/calculate',
+  serverLoggingUrl: 'https://glimmpse.samplesizeshop.org/api/clientsidelog',
+  calculateUrl: 'https://glimmpse.samplesizeshop.org/api/calculate',
   loglevel: NgxLoggerLevel.DEBUG,
   serverLoglevel: NgxLoggerLevel.WARN,
   disableAnimations: false,
